modified:   luogu/p1024一元三次方程求解.cpp 	new file:   luogu/p1536村村通.cpp 	new file:   luogu/p1725琪露诺.cpp 	modified:   luogu/p1886[模板]单调队列.cpp 	new file:   luogu/p2422良好的感觉.cpp 	new file:   luogu/p2629好消息坏消息.cpp 	new file:   luogu/p2947_Look_up_S.cpp 	new file:   luogu/p5250木材仓库.cpp 	modified:   template/单调队列.cpp

+++ luogu/p1024一元三次方程求解.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 using namespace std;
 
-int a, b, c, d;
+double a, b, c, d;
 
 double func(double);
 void bin(double l, double r)
